# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/events/{id}":
    get:
      tags:
        - Events
      summary: Получить мероприятие по ID
      operationId: eventsIdGet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Мероприятие
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventResponse"
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Events
      summary: Обновить мероприятие (Admin)
      operationId: eventsIdPut
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventUpdateRequest"
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventResponse"
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Конфликт
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
    delete:
      tags:
        - Events
      summary: Удалить мероприятие (Admin)
      operationId: eventsIdDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Мероприятие удалено
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Конфликт
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
...
  "/events":
    get:
      tags:
        - Events
      summary: Получить список мероприятий
      operationId: eventsGet
      parameters:
        - name: from
          in: query
          description: Фильтр от даты
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Фильтр до даты
          required: false
          schema:
            type: string
            format: date-time
        - name: pageable
          in: query
          description: Параметры пагинации и сортировки
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: Список мероприятий
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageableEventResponse"
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Events
      summary: Создать мероприятие (Admin)
      operationId: eventsPost
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventCreateRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventResponse"
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Конфликт
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
  "/events/{id}/cover":
    post:
      tags:
        - Events
      summary: Загрузить обложку мероприятия (Admin)
      operationId: eventsIdCoverPost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cover:
                  type: string
                  description: Файл изображения (JPEG/PNG, макс. 5MB)
                  format: binary
      responses:
        '200':
          description: Обложка обновлена
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventResponse"
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Конфликт
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '413':
          description: Файл слишком большой
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
    delete:
      tags:
        - Events
      summary: Удалить обложку мероприятия (Admin)
      operationId: eventsIdCoverDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Обложка удалена
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
...
components:
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Должен быть валидным email адресом
      description: Детали ошибки (опционально)
    ErrorResponse:
      required:
        - code
        - level
        - message
      type: object
      properties:
        code:
          type: string
          description: Машинно-читаемый код ошибки
          example: VALIDATION_FAILED
        level:
          type: string
          description: Уровень серьезности
          example: error
          enum:
            - ErrorResponse.LevelEnum.INFO(value=info)
            - ErrorResponse.LevelEnum.WARNING(value=warning)
            - ErrorResponse.LevelEnum.ERROR(value=error)
        message:
          type: string
          description: Человеко-читаемое сообщение
          example: Некоторые поля заполнены неверно
        details:
          type: array
          description: Детали ошибки (опционально)
          items:
            "$ref": "#/components/schemas/ErrorDetail"
...
    EventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        dateTime:
          type: string
          format: date-time
        totalTickets:
          type: integer
          format: int32
        availableTickets:
          type: integer
          format: int32
        coverUrl:
          type: string
          description: URL обложки мероприятия
          example: "/uploads/events/1/cover.jpg"
    EventUpdateRequest:
      type: object
      properties:
        title:
          maxLength: 100
          minLength: 0
          type: string
        description:
          type: string
        dateTime:
          type: string
          format: date-time
        totalTickets:
          minimum: 1
          type: integer
          format: int32
...
    EventCreateRequest:
      required:
        - dateTime
        - title
        - totalTickets
      type: object
      properties:
        title:
          maxLength: 100
          minLength: 0
          type: string
        description:
          type: string
        dateTime:
          type: string
          format: date-time
        totalTickets:
          minimum: 1
          type: integer
          format: int32
...
    PageableEventResponse:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/EventResponse"
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            "$ref": "#/components/schemas/SortObject"
        first:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
      description: Страница с мероприятиями
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            "$ref": "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...